Initial Code Commit
This changeset includes my default set of configuration files for any
go library/cli. This includes a set of pre-commit hooks for developer
validation prior to commiting code as well as a set of github actions
for CI/CD and release management.

These configurations also include support for a full set of Go linters
using the Golang CI lint tool and a dependabot configuration for
ensuring any dependencies for either the go.mod or github actions are
up to date.