Library Implementation Commit
This commit includes the initial library implementation for the exercise
as well as a number of helper binaries for testing and isolation.

The usage of this library is detailed in the README.md and the docs/design.md.

There were several adjustments that were necessary to the original design and
API once I started implementation. These have also been updated in the design
document.